requirements.txt
bencoder.pyx
redis

